refactor: problemTest => test로 폴더명 수정
Related to: #22
