+++ containers/problems/index.tsx
@@ -1,4 +1,4 @@
 export { default as ProblemTemplate } from './ProblemTemplate';
-export { default as ProblemTest } from './problemTest';
+export { default as ProblemTest } from './test';
 export * from './code';
 export * from './description';
